Moving DTOs to Request namespace
