+++ ShoeTracker/ShoeTracker.Server/Controllers/ShoeController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ShoeTracker.Server.DataAccess.Models;
-using ShoeTracker.Server.Models;
+using ShoeTracker.Server.Models.Request;
 using ShoeTracker.Server.Service;
 
 namespace ShoeTracker.Server.Controllers
