+++ ShoeTracker/ShoeTracker.Server/Models/Request/CreateActivityDto.cs
@@ -1,4 +1,4 @@
-﻿namespace ShoeTracker.Server.Models
+﻿namespace ShoeTracker.Server.Models.Request
 {
     public class CreateActivityDto
     {
