+++ ShoeTracker/ShoeTracker.Server/Models/Request/CreateShoeDto.cs
@@ -1,21 +1,18 @@
-﻿namespace ShoeTracker.Server.Models
+﻿namespace ShoeTracker.Server.Models.Request
 {
-    public class GetShoeDto
+    public class CreateShoeDto
     {
-        public Guid Id { get; set; }
-
         public string Brand { get; set; } = string.Empty;
 
         public string Model { get; set; } = string.Empty;
 
-        // 0 indicates missing value, should always be >=1
         public uint ModelVersion { get; set; }
 
         public string ShoeName { get; set; } = string.Empty;
 
         public string? Description { get; set; } = null;
 
-        public double Miles { get; set; }
+        public double StartingMiles { get; set; }
 
         public TextColor TextColor { get; set; }
 
@@ -24,7 +21,5 @@
         public DateModel StartDate { get; set; }
 
         public double WarnAtMileage { get; set; }
-
-        public double StartingMileage { get; set; }
     }
 }
