+++ ShoeTracker/ShoeTracker.Server/Models/Request/GetShoeDto.cs
@@ -1,18 +1,21 @@
-﻿namespace ShoeTracker.Server.Models
+﻿namespace ShoeTracker.Server.Models.Request
 {
-    public class CreateShoeDto
+    public class GetShoeDto
     {
+        public Guid Id { get; set; }
+
         public string Brand { get; set; } = string.Empty;
 
         public string Model { get; set; } = string.Empty;
 
+        // 0 indicates missing value, should always be >=1
         public uint ModelVersion { get; set; }
 
         public string ShoeName { get; set; } = string.Empty;
 
         public string? Description { get; set; } = null;
 
-        public double StartingMiles { get; set; }
+        public double Miles { get; set; }
 
         public TextColor TextColor { get; set; }
 
@@ -21,5 +24,7 @@
         public DateModel StartDate { get; set; }
 
         public double WarnAtMileage { get; set; }
+
+        public double StartingMileage { get; set; }
     }
 }
